requirements.txt
opencv-python>=4.5.0
python-telegram-bot>=13.0
python-dotenv>=0.19.0
schedule>=1.1.0
